Rename changedetector and imagerestorer
